requirements.txt
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1636012310340/work
certifi==2021.10.8
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/croot-2p6q7e2r/cffi_1642701115063/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1639699399793/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1635862551662/work
PyYAML==6.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
spotipy @ file:///home/conda/feedstock_root/build_artifacts/spotipy_1631215603968/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1641584929973/work
